Flatten optimize module structure by moving lagrange/ contents up one level
Moved core.py, derivative.py, and minimize.py from lagrange/ subdirectory to
main optimize/ directory. Updated import paths in source and test files.
Removed empty lagrange/ directory structure. All tests pass.
