+++ src/exogibbs/optimize/core.py
