+++ src/exogibbs/optimize/derivative.py
