+++ tests/unittests/optimize/lagrange/minimize_vmap_test.py
@@ -1,8 +1,8 @@
 import pytest
 import jax.numpy as jnp
 from jax import config, vmap, jacrev
-from exogibbs.optimize.lagrange.minimize import minimize_gibbs
-from exogibbs.optimize.lagrange.core import compute_ln_normalized_pressure
+from exogibbs.optimize.minimize import minimize_gibbs
+from exogibbs.optimize.core import compute_ln_normalized_pressure
 from exogibbs.test.analytic_hsystem import HSystem
